favorites api in product controller
