--- src/main/resources/http/shop/favorites.http
@@ -1,24 +1,0 @@
-### Favorites List
-GET http://localhost:8080/favorites?userId={{userId}}
-Content-Type: application/json
-Authorization: Bearer {{accessToken}}
-
-### Favorites Add
-POST http://localhost:8080/favorites
-Content-Type: application/json
-Authorization: Bearer {{accessToken}}
-
-{
-  "userId": {{userId}},
-  "productId": {{productId}}
-}
-
-### Favorites Remove
-DELETE http://localhost:8080/favorites
-Content-Type: application/json
-Authorization: Bearer {{accessToken}}
-
-{
-  "userId": {{userId}},
-  "productId": {{productId}}
-}
