+++ src/main/resources/http/shop/product.http
@@ -99,3 +99,18 @@ Authorization: Bearer {{accessToken}}
 DELETE http://localhost:8080/products/{{productId}}
 Content-Type: application/json
 Authorization: Bearer {{accessToken}}
+
+### Favorites List
+GET http://localhost:8080/favorites
+Content-Type: application/json
+Authorization: Bearer {{accessToken}}
+
+### Favorites Add
+POST http://localhost:8080/products/{{productId}}/add-to-favorites
+Content-Type: application/json
+Authorization: Bearer {{accessToken}}
+
+### Favorites Remove
+DELETE http://localhost:8080/products/{{productId}}/remove-from-favorites
+Content-Type: application/json
+Authorization: Bearer {{accessToken}}
